Extract window resources, separate enums
