+++ AutoClicker/MainWindow.xaml
@@ -2,41 +2,11 @@
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:local="clr-namespace:AutoClicker"
-        xmlns:system="clr-namespace:System;assembly=mscorlib"
+        xmlns:enums="clr-namespace:AutoClicker.Enums"
         ResizeMode="CanMinimize"
         Height="320" Width="450">
-
     <Window.Resources>
-        <ObjectDataProvider x:Key="mouseButtonValues"
-                            MethodName="GetValues"
-                            ObjectType="{x:Type system:Enum}">
-            <ObjectDataProvider.MethodParameters>
-                <x:Type TypeName="local:MouseButton"/>
-            </ObjectDataProvider.MethodParameters>
-        </ObjectDataProvider>
-        <ObjectDataProvider x:Key="mouseActionValues"
-                            MethodName="GetValues"
-                            ObjectType="{x:Type system:Enum}">
-            <ObjectDataProvider.MethodParameters>
-                <x:Type TypeName="local:MouseAction"/>
-            </ObjectDataProvider.MethodParameters>
-        </ObjectDataProvider>
-        <ObjectDataProvider x:Key="repeatModeValues"
-                            MethodName="GetValues"
-                            ObjectType="{x:Type system:Enum}">
-            <ObjectDataProvider.MethodParameters>
-                <x:Type TypeName="local:RepeatMode"/>
-            </ObjectDataProvider.MethodParameters>
-        </ObjectDataProvider>
-        <ObjectDataProvider x:Key="locationModeValues"
-                            MethodName="GetValues"
-                            ObjectType="{x:Type system:Enum}">
-            <ObjectDataProvider.MethodParameters>
-                <x:Type TypeName="local:LocationMode"/>
-            </ObjectDataProvider.MethodParameters>
-        </ObjectDataProvider>
-        <local:EnumToBooleanConverter x:Key="EnumToBooleanConverter" />
-        <BitmapImage x:Key="HelpIcon" UriSource="Resources/question.png" />
+        <ResourceDictionary Source="Resources/ResourceDictionary.xaml"/>
     </Window.Resources>
 
     <Window.CommandBindings>
@@ -122,10 +92,10 @@
                     <RadioButton Grid.Row="0" Grid.Column="0" Margin="0, 0, 0, 5"
                                  Content="Infinite (Until stopped)"
                                  IsChecked="{Binding Path=SelectedRepeatMode, Converter={StaticResource EnumToBooleanConverter},
-                                        ConverterParameter={x:Static local:RepeatMode.Infinite}}"/>
+                                        ConverterParameter={x:Static enums:RepeatMode.Infinite}}"/>
                     <RadioButton Grid.Row="1" Grid.ColumnSpan="2" Margin="0, 0, 0, 5"
                                  IsChecked="{Binding Path=SelectedRepeatMode, Converter={StaticResource EnumToBooleanConverter},
-                                        ConverterParameter={x:Static local:RepeatMode.Count}}">
+                                        ConverterParameter={x:Static enums:RepeatMode.Count}}">
                         <StackPanel Orientation="Horizontal">
                             <TextBox Width="50" Margin="5, 0, 0, 0"
                                      Text="{Binding SelectedTimesToRepeat, UpdateSourceTrigger=PropertyChanged, Mode=TwoWay}"/>
@@ -149,10 +119,10 @@
                     <RadioButton Grid.Row="0" Grid.Column="0" Margin="0, 0, 0, 5"
                              Content="Current Cursor Position"
                              IsChecked="{Binding Path=SelectedLocationMode, Converter={StaticResource EnumToBooleanConverter},
-                                        ConverterParameter={x:Static local:LocationMode.CurrentLocation}}"/>
+                                        ConverterParameter={x:Static enums:LocationMode.CurrentLocation}}"/>
                     <RadioButton Grid.Row="1" Grid.ColumnSpan="2" Margin="0, 0, 0, 5"
                              IsChecked="{Binding Path=SelectedLocationMode, Converter={StaticResource EnumToBooleanConverter},
-                                        ConverterParameter={x:Static local:LocationMode.PickedLocation}}">
+                                        ConverterParameter={x:Static enums:LocationMode.PickedLocation}}">
                         <StackPanel Orientation="Horizontal">
                             <TextBlock Text="X" Margin="5, 0, 0, 0"/>
                             <TextBox Width="50" Margin="5, 0, 0, 0"
